requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.7.0
billiard==4.1.0
black==23.3.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==2.0.0
dj-email-url==1.0.6
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cache-url==3.4.4
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-spectacular==0.26.3
environs==9.5.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kombu==5.3.1
marshmallow==3.19.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.8.8
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
